updates seed script (types and extracts countries.js)
