+++ api/prisma/countries.js
@@ -1,0 +1,64 @@
+const countries = [
+  {
+    iso: 'itl',
+    worldometersSlug: 'italy',
+    name: 'Italy',
+  },
+  {
+    iso: 'chn',
+    worldometersSlug: 'china',
+    name: 'China',
+  },
+  {
+    iso: 'irn',
+    worldometersSlug: 'iran',
+    name: 'Iran',
+  },
+  {
+    iso: 'kor',
+    worldometersSlug: 'south-korea',
+    name: 'South Korea',
+  },
+  {
+    iso: 'esp',
+    worldometersSlug: 'spain',
+    name: 'Spain',
+  },
+  {
+    iso: 'deu',
+    worldometersSlug: 'germany',
+    name: 'Germany',
+  },
+  {
+    iso: 'fra',
+    worldometersSlug: 'france',
+    name: 'France',
+  },
+  {
+    iso: 'usa',
+    worldometersSlug: 'us',
+    name: 'United States',
+  },
+  {
+    iso: 'gbr',
+    worldometersSlug: 'uk',
+    name: 'United Kingdom',
+  },
+  {
+    iso: 'cad',
+    worldometersSlug: 'canada',
+    name: 'Canada',
+  },
+  {
+    iso: 'ind',
+    worldometersSlug: 'india',
+    name: 'India',
+  },
+  {
+    iso: 'bra',
+    worldometersSlug: 'brazil',
+    name: 'Brazil',
+  },
+]
+
+module.exports = countries
