+++ api/prisma/seed.js
@@ -1,77 +1,22 @@
 /* eslint-disable no-console */
 const { PrismaClient } = require('@prisma/client')
 const dotenv = require('dotenv')
+const countries = require('./countries')
 
 dotenv.config()
 const db = new PrismaClient()
 
-const countries = [
-  {
-    iso: 'itl',
-    worldometersSlug: 'italy',
-    name: 'Italy',
-  },
-  {
-    iso: 'chn',
-    worldometersSlug: 'china',
-    name: 'China',
-  },
-  {
-    iso: 'irn',
-    worldometersSlug: 'iran',
-    name: 'Iran',
-  },
-  {
-    iso: 'kor',
-    worldometersSlug: 'south-korea',
-    name: 'South Korea',
-  },
-  {
-    iso: 'esp',
-    worldometersSlug: 'spain',
-    name: 'Spain',
-  },
-  {
-    iso: 'deu',
-    worldometersSlug: 'germany',
-    name: 'Germany',
-  },
-  {
-    iso: 'fra',
-    worldometersSlug: 'france',
-    name: 'France',
-  },
-  {
-    iso: 'usa',
-    worldometersSlug: 'us',
-    name: 'United States',
-  },
-  {
-    iso: 'gbr',
-    worldometersSlug: 'uk',
-    name: 'United Kingdom',
-  },
-  // {
-  //   iso: 'cad',
-  //   worldometersSlug: 'canada',
-  //   name: 'Canada',
-  // },
-  // {
-  //   iso: 'ind',
-  //   worldometersSlug: 'india',
-  //   name: 'India',
-  // },
-  // {
-  //   iso: 'bra',
-  //   worldometersSlug: 'brazil',
-  //   name: 'Brazil',
-  // },
-]
-
+/**
+ *
+ * @template T
+ * @param {Array<T>} array
+ * @param {(item: T) => void} callback
+ * @returns Promise<void>
+ */
 async function asyncForEach(array, callback) {
   const promises = array.map((item) => callback(item))
 
-  return Promise.all(promises)
+  return Promise.allSettled(promises)
 }
 
 async function main() {
